requirements.txt
anyio==3.4.0
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
distlib==0.3.4
Django==4.0
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-multiselectfield==0.1.12
django-notifications-hq==1.7.0
fastapi==0.70.0
filelock==3.7.1
h11==0.12.0
hyperlink==21.0.0
idna==3.3
image==1.5.33
incremental==21.3.0
jsonfield==3.1.0
listener==2.0.0a1
logging==0.4.9.6
MouseInfo==0.1.3
mysql-connector-python==8.0.29
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
protobuf==4.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/shawo/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyOpenSSL==22.1.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.4
pywin32==302
qrcode==7.3.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
SpeechRecognition==3.8.1
sqlparse==0.4.2
starlette==0.16.0
style==1.1.0
swapper==1.3.0
toml==0.10.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.0.0
tzdata==2021.5
update==0.0.1
urllib3==1.26.12
uvicorn==0.15.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
zope.interface==5.4.0
